ffff

+++ mirrord/agent/src/file.rs
@@ -299,7 +299,7 @@ impl FileManager {
                             // Limit the new position to `buffer_size`.
                             Ok((
                                 read_amount,
-                                position_after_read.clamp(0, buffer_size as u64),
+                                position_after_read.clamp(0, buffer_size),
                             ))
                         })
                         .and_then(|(read_amount, seek_to)| {
